requirements.txt
aws-cdk-lib==2.114.1
constructs>=10.0.0,<11.0.0
boto3>=1.34.0
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.88.0